requirements.txt
langchain-openai
langchain-community
langchain-chroma
langchain-core
langchain-google-genai

beautifulsoup4
google

requests

llmsherpa
llama-index

pypdf

streamlit